# packages/contracts/openapi/storage-console-v1.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "Counts": {
        "additionalProperties": false,
        "properties": {
          "filesystem_objects": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Filesystem Objects",
            "type": "integer"
          },
          "shares": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Shares",
            "type": "integer"
          },
          "sources": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Sources",
            "type": "integer"
          },
          "volumes": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Volumes",
            "type": "integer"
          }
        },
        "required": [
          "sources",
          "volumes",
          "shares",
          "filesystem_objects"
        ],
        "title": "Counts",
        "type": "object"
      },
...
      "Freshness": {
        "additionalProperties": false,
        "properties": {
          "age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Seconds"
          },
          "bottleneck_collector_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bottleneck Collector Id"
          },
          "collector_count": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Collector Count",
            "type": "integer"
          },
          "cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor"
          },
          "expected_cadence_seconds": {
            "title": "Expected Cadence Seconds",
            "type": "integer"
          },
          "lag_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lag Seconds"
          },
          "last_collector_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Collector At"
          },
          "last_event_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event At"
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "state": {
            "enum": [
              "HEALTHY",
              "OBSERVE",
              "WARNING",
              "CRITICAL",
              "UNKNOWN",
              "NOT_APPLICABLE"
            ],
            "title": "State",
            "type": "string"
          },
          "unknown_collector_count": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Unknown Collector Count",
            "type": "integer"
          }
        },
        "required": [
          "state",
          "reason",
          "expected_cadence_seconds",
          "last_success_at",
          "last_event_at",
          "last_collector_at",
          "age_seconds",
          "cursor",
          "lag_seconds",
          "collector_count",
          "unknown_collector_count",
          "bottleneck_collector_id"
        ],
        "title": "Freshness",
        "type": "object"
      },
...
      "Overview": {
        "additionalProperties": false,
        "properties": {
          "counts": {
            "$ref": "#/components/schemas/Counts"
          },
          "domains": {
            "items": {
              "$ref": "#/components/schemas/DomainHealth"
            },
            "title": "Domains",
            "type": "array"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessSummary"
          },
          "overall_state": {
            "enum": [
              "HEALTHY",
              "OBSERVE",
              "WARNING",
              "CRITICAL",
              "UNKNOWN",
              "NOT_APPLICABLE"
            ],
            "title": "Overall State",
            "type": "string"
          }
        },
        "required": [
          "domains",
          "evaluated_at",
          "counts",
          "overall_state",
          "freshness"
        ],
        "title": "Overview",
        "type": "object"
      },
      "Page_Share_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Share"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Share]",
        "type": "object"
      },
      "Page_Source_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Source]",
        "type": "object"
      },
      "Page_Volume_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Volume"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[Volume]",
        "type": "object"
      },
...
      "Source": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fqdn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fqdn"
          },
          "freshness": {
            "$ref": "#/components/schemas/Freshness"
          },
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instance_id": {
            "title": "Instance Id",
            "type": "string"
          },
          "source_type": {
            "enum": [
              "FILESERVER",
              "PVE",
              "PBS"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source_type",
          "hostname",
          "fqdn",
          "instance_id",
          "created_at",
          "freshness"
        ],
        "title": "Source",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CollectorToken": {
        "scheme": "bearer",
        "type": "http"
      },
      "UserSession": {
        "in": "cookie",
        "name": "__Host-storage_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Storage Console API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/overview": {
      "get": {
        "operationId": "overview_api_v1_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Evidence-Valid-For-Ms": {
                "description": "Relative evidence lifetime from snapshot time; clients subtract complete request elapsed time.",
                "schema": {
                  "maximum": 35000.0,
                  "minimum": 0.0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "read"
        ]
      }
    },
    "/api/v1/shares": {
      "get": {
        "operationId": "share_list_api_v1_shares_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Share_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Evidence-Valid-For-Ms": {
                "description": "Relative evidence lifetime from snapshot time; clients subtract complete request elapsed time.",
                "schema": {
                  "maximum": 35000,
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Share List",
        "tags": [
          "read"
        ]
      }
    },
    "/api/v1/sources": {
      "get": {
        "operationId": "sources_api_v1_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Source_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Evidence-Valid-For-Ms": {
                "description": "Relative evidence lifetime from snapshot time; clients subtract complete request elapsed time.",
                "schema": {
                  "maximum": 35000,
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Sources",
        "tags": [
          "read"
        ]
      }
    },
    "/api/v1/sources/{source_id}": {
      "get": {
        "operationId": "source_detail_api_v1_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Evidence-Valid-For-Ms": {
                "description": "Relative evidence lifetime from snapshot time; clients subtract complete request elapsed time.",
                "schema": {
                  "maximum": 35000,
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Source Detail",
        "tags": [
          "read"
        ]
      }
    },
    "/api/v1/sources/{source_id}/freshness": {
      "get": {
        "operationId": "source_freshness_api_v1_sources__source_id__freshness_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Freshness"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Evidence-Valid-For-Ms": {
                "description": "Relative evidence lifetime from snapshot time; clients subtract complete request elapsed time.",
                "schema": {
                  "maximum": 35000,
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Source Freshness",
        "tags": [
          "read"
        ]
      }
    },
    "/api/v1/volumes": {
      "get": {
        "operationId": "volume_list_api_v1_volumes_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Volume_"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Evidence-Valid-For-Ms": {
                "description": "Relative evidence lifetime from snapshot time; clients subtract complete request elapsed time.",
                "schema": {
                  "maximum": 35000,
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "UserSession": []
          }
        ],
        "summary": "Volume List",
        "tags": [
          "read"
        ]
      }
    },
...
  }
}